requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-generativeai==0.8.0
pydantic==2.7.4
pydantic-settings==2.4.0
pytest==7.4.3
httpx==0.27.0
PyJWT==2.8.0
requests==2.31.0
